update to a plate demo
